requirements.txt
python-dotenv
pypdf2
openai
chromadb
langchain

